Add Promise Support & Refactor Module to ES6
